kissplice 2.6.5 (new formula)
Kissplice was packaged before the migration to brewsci/bio :
https://github.com/brewsci/homebrew-science/blob/master/Formula/kissplice.rb

Modernize packaging with the current configuration :
- gitlab
- modern cmake
- run sample example as test stage. another possibility is running make test in install

Kissplice depends on bcalm, but uses an internal build that is patched to work with ARM MacOS.
The patch has been submitted to bcalm upstream but in the meantime the current bcalm formula does not build on ARM MacOS.

Added patches to remove compiler path (with shim directory) from bcalm.
These patches are quite ugly but necessary to pass brew audit, and would be needed in thebcalm formula too.